Improving Plugin Handling in Our System
We have made an improvement to the script that handles plugins in our system.
Previously, if a directory necessary for a plugin was not present, the script
would skip it. However, we have discovered that the correct approach is to
create the necessary directories at the base level and create a symlink of the
plugin config there. To implement this change, we have added the "mkpath"
function from the "File::Path" module and used it to create the necessary
directories if they do not already exist. This update will ensure that the
plugin handling process runs smoothly and efficiently in the future.