Adding Git URL Output to Plugin Script
As part of the development process for a new plugin, the team decided to add
some additional logging to the plugin's installation script. Specifically, they
wanted to output the Git URL of the repository that the plugin was being pulled
from. This would be useful for both the user and the developer, as it would
provide a clear indication of where the plugin's source code was being obtained
from.

To implement this change, the team made a small modification to the plugin
script, adding a single line of code to output the Git URL to the console. They
then tested the updated script to ensure that it was functioning correctly and
that the Git URL was being displayed as expected. Overall, this was a simple
but useful addition to the plugin script that provided valuable information to
both users and developers