Eliminating Silly Repetitive Tasks
As a developer, we all know the frustration of having to run the same command
multiple times. It's a waste of time and energy, and it can be especially
frustrating when it's something as simple as updating a plugin.

That's where this patch comes in. When the plugin update script is called, it
will now check if it has been updated itself. If it has, it will automatically
run the update command again without the user having to do it manually.

This may seem like a small change, but it can make a big difference in
streamlining your workflow and eliminating unnecessary steps. So next time
you're updating your plugin, you won't have to worry about running the command
twice - the script will take care of it for you.

This patch is just one example of how small changes can have a big impact, and
it's a reminder that even the smallest improvements can make a difference in
our daily work.